Split out primary and secondary index building/access
