+++ pkg/indexbuilder/ottawau.py
@@ -7,12 +7,14 @@ from collections import defaultdict
 import nltk
 from .indexbuilder import IndexBuilder
 from .invertedindex import IndexValue
+from ..indexaccess import IndexAccessor
 from ..dictionary import Dictionary, DictBuilder
 from ..wordmodifiers import context
 
 
 class OttawaUIndexBuilder(IndexBuilder):
     def __init__(self, ctx):
+        self.ctx = ctx
         self.corpus_path = ctx.corpus_path
         self.dict_path = ctx.dict_path
         self.inverted_index_path = ctx.inverted_index_path
@@ -32,6 +34,31 @@ class OttawaUIndexBuilder(IndexBuilder):
             yaml.dump(
                 inverted_index,
                 index_file,
+                explicit_start=True,
+                default_flow_style=True,
+                sort_keys=False,
+                indent=2,
+            )
+
+    def build_bigram_index(self):
+        index = defaultdict(default_index_value)
+
+        # This is a hacky way to get access to all the keys in a given index
+        # In general, we only expose 'get' access to a single key at a time, not the full set of keys.
+        # If we have time, we can refactor into something nicer but at the end of the day
+        # it's just going to be doing this so I'm not too worried.
+        index_accessor = IndexAccessor(self.ctx)
+        keys = index_accessor.index[self.ctx.inverted_index_path].index.keys()
+
+        for key in keys:
+            k = f"${key}$"  # add begin/end indicators
+            # SOURCE: https://stackoverflow.com/questions/21303224/iterate-over-all-pairs-of-consecutive-items-in-a-list
+            for first, second in zip(k, k[1:]):
+                index[first + second].append(key)
+        with open(self.ctx.bigram_index_path(), "w") as bigram_handle:
+            yaml.dump(
+                index,
+                bigram_handle,
                 explicit_start=True,
                 default_flow_style=True,
                 sort_keys=False,
@@ -58,3 +85,7 @@ class OttawaUIndexBuilder(IndexBuilder):
                     if dictionary.contains(term):
                         simple_index[term].append(document.id)
         return simple_index
+
+
+def default_index_value():
+    return []
